+++ src/blueprints/telegram_bot/webhook/commands/create_folder.py
@@ -1,4 +1,4 @@
-from flask import g
+from flask import g, current_app
 
 from src.api import telegram
 from src.blueprints.telegram_bot._common.yandex_disk import (
@@ -6,39 +6,61 @@ from src.blueprints.telegram_bot._common.yandex_disk import (
     YandexAPICreateFolderError,
     YandexAPIRequestError
 )
-from ._common.responses import (
-    cancel_command,
-    abort_command,
-    AbortReason
+from src.blueprints.telegram_bot._common.command_names import (
+    CommandName
 )
+from src.blueprints.telegram_bot._common.stateful_chat import (
+    set_disposable_handler
+)
+from src.blueprints.telegram_bot.webhook.dispatcher_events import (
+    DispatcherEvent,
+    RouteSource
+)
+from ._common.responses import cancel_command
 from ._common.decorators import (
     yd_access_token_required,
     get_db_data
 )
-from src.blueprints.telegram_bot._common.command_names import CommandName
 
 
 @yd_access_token_required
 @get_db_data
 def handle(*args, **kwargs):
-    """
-    Handles `/create_folder` command.
-    """
-    message = g.telegram_message
-    user = g.db_user
-    chat = g.db_chat
-    message_text = message.get_text()
-    folder_name = message_text.replace(
-        CommandName.CREATE_FOLDER.value,
-        ""
-    ).strip()
+    message = kwargs.get(
+        "message",
+        g.telegram_message
+    )
+    user_id = kwargs.get(
+        "user_id",
+        g.telegram_user.id
+    )
+    chat_id = kwargs.get(
+        "chat_id",
+        g.telegram_chat.id
+    )
+    folder_name = get_folder_name(
+        message,
+        kwargs.get("route_source")
+    )
 
-    if not (folder_name):
-        return abort_command(
-            chat.telegram_id,
-            AbortReason.NO_SUITABLE_DATA
+    if not folder_name:
+        set_disposable_handler(
+            user_id,
+            chat_id,
+            CommandName.CREATE_FOLDER.value,
+            [
+                DispatcherEvent.PLAIN_TEXT.value,
+                DispatcherEvent.BOT_COMMAND.value,
+                DispatcherEvent.EMAIL.value,
+                DispatcherEvent.HASHTAG.value,
+                DispatcherEvent.URL.value
+            ],
+            current_app.config["RUNTIME_DISPOSABLE_HANDLER_EXPIRE"]
         )
 
+        return message_wait_for_text(chat_id)
+
+    user = g.db_user
     access_token = user.yandex_disk_token.get_access_token()
     last_status_code = None
 
@@ -48,19 +70,13 @@ def handle(*args, **kwargs):
             folder_name=folder_name
         )
     except YandexAPIRequestError as error:
-        print(error)
-        return cancel_command(chat.telegram_id)
+        cancel_command(chat_id)
+        raise error
     except YandexAPICreateFolderError as error:
-        error_text = (
-            str(error) or
-            "Unknown Yandex.Disk error"
-        )
+        message_yandex_error(chat_id, str(error))
 
-        telegram.send_message(
-            chat_id=chat.telegram_id,
-            text=error_text
-        )
-
+        # it is expected error and should be
+        # logged only to user
         return
 
     text = None
@@ -70,9 +86,54 @@ def handle(*args, **kwargs):
     elif (last_status_code == 409):
         text = "Already exists"
     else:
-        text = f"Unknown status code: {last_status_code}"
+        text = f"Unknown operation status: {last_status_code}"
 
     telegram.send_message(
-        chat_id=chat.telegram_id,
+        chat_id=chat_id,
         text=text
+    )
+
+
+def get_folder_name(
+    telegram_message,
+    route_source: RouteSource
+) -> str:
+    folder_name = telegram_message.get_text()
+
+    # On "Disposable handler" route we expect pure text,
+    # in other cases we expect bot command as start of a message
+    if (route_source != RouteSource.DISPOSABLE_HANDLER):
+        folder_name = folder_name.replace(
+            CommandName.CREATE_FOLDER.value,
+            "",
+            1
+        ).strip()
+
+    return folder_name
+
+
+def message_wait_for_text(chat_id: int) -> None:
+    telegram.send_message(
+        chat_id=chat_id,
+        parse_mode="HTML",
+        text=(
+            "Send a folder name to create."
+            "\n\n"
+            "It should starts from root directory, "
+            "nested folders should be separated with "
+            '"<code>/</code>" character. '
+            "In short, i expect a full path."
+            "\n\n"
+            "Example: <code>Telegram Bot/kittens and raccoons</code>"
+        )
+    )
+
+
+def message_yandex_error(chat_id: int, error_text: str) -> None:
+    telegram.send_message(
+        chat_id=chat_id,
+        text=(
+            error_text or
+            "Unknown Yandex.Disk error"
+        )
     )